+++ tradin/app/backtest/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import BackTestPage from "../../src/components/units/backtest/BackTest";
+import BackTestPage from "../../src/components/units/backtest/main/BackTest";
 
 export default function BackTest() {
     return (
