+++ tradin/src/components/units/backtest/mockdata/MockStrategy.tsx
