+++ tradin/src/components/units/backtest/result/BackTestResultSkeletonUI.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from "@emotion/styled";
 import { keyframes } from "@emotion/react";
-import * as S from './BackTest.styles';
+import * as S from '../main/BackTest.styles';
 
 const shimmer = keyframes`
     0% { background-position: 200% 0; }
