+++ tradin/src/components/units/backtest/result/CalculateMetrics.tsx
